requirements.txt
streamlit
streamlit-authenticator
gspread
google-auth
pandas
requests
bcrypt
python-dateutil
pytz
